requirement.txt
pandas
numpy
json5  
requests
streamlit
faiss-cpu
sentence-transformers
redis
